feat: Export createCompiledCatalog to public api
